requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.1
